Fix the more-than-two-argument case for `list='.
This fix is due to Peter Lane and Per Bothner.  Thanks to John Cowan
and Taylor Campbell for suggestions.

I've modified the change from Per's patch to preserve tabs and
whitespace, to make comparing versions easier, and to eliminate the
renaming of `=' to `elt=', which was not necessary for the fix.