Fix implementation: arguments to = were flipped.
It's an equality predicate, but the document specified the order in
which they were to be passed, and the sample implementation didn't
match that.  This fixes the problem.

Thanks to Chris Hanson for reporting the bug and to Alex Shinn for
help refining the fix.

See this email thread for the initial report and ensuing discussion:

  https://srfi-email.schemers.org/srfi-1/msg/14189904/